refactor: remove placeholder data paths
